Consolidate data files, create csvs from which df can be made.
prechart now looks in a directory of ignition txt files. It consolidates all the
data into a single txt file. That single file gets turned into a csv file from
which the prechart df can be made.

Rename main.py --> summarize.py. main is going to be the file that both
summarizes and makes prechart.